+++ ft_void_ptr.c
@@ -6,13 +6,13 @@
 /*   By: gkrusta <gkrusta@student.42malaga.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/04 12:21:26 by gkrusta           #+#    #+#             */
-/*   Updated: 2023/05/04 13:27:28 by gkrusta          ###   ########.fr       */
+/*   Updated: 2023/05/05 13:44:52 by gkrusta          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "printf.h"
+#include "ft_printf.h"
 
-int	ft_void_ptr(unsigned long n)
+int	ft_void_ptr(unsigned long long n)
 {
 	int	i;
